Move club router to package api/club
